Move all service files to new service folder and adjust affected files accordingly.
