+++ src/app/specs/disable-alias-bttn.service.spec.ts
@@ -1,5 +1,5 @@
 import { async, inject, TestBed } from "@angular/core/testing";
-import { DisableAliasBttnService } from "../disable-alias-bttn.service";
+import { DisableAliasBttnService } from "../services/disable-alias-bttn.service";
 
 describe("DisableAliasBttnService", () => {
   let service: DisableAliasBttnService;
