+++ src/app/specs/metas.component.spec.ts
@@ -1,11 +1,10 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, DebugElement } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { NO_ERRORS_SCHEMA, DebugElement } from "@angular/core";
+import { HttpModule } from "@angular/http";
+import { By } from "@angular/platform-browser";
 
-import { By } from '@angular/platform-browser';
-
-import { MetasComponent } from '../components/metas.component';
-import { MetaService } from '../meta-service';
+import { MetasComponent } from "../components/metas.component";
+import { MetaService } from "../services/meta-service";
 
 let comp: MetasComponent;
 let fixture: ComponentFixture<MetasComponent>;
@@ -33,17 +32,17 @@ describe("MetasComponent", () => {
     fixture = TestBed.createComponent(MetasComponent);
     metaService = fixture.debugElement.injector.get(MetaService);
     DOMElement = fixture.nativeElement.children;
-    serviceSpy = spyOn(metaService, 'getMetas').and.returnValue(Promise.resolve(0));
+    serviceSpy = spyOn(metaService, "getMetas").and.returnValue(Promise.resolve(0));
     comp = fixture.componentInstance;
     fixture.detectChanges();
   });
 
 
-  it('ngOnInit() should call the sevice method function', () => {
+  it("ngOnInit() should call the sevice method function", () => {
     expect(serviceSpy).toHaveBeenCalled();
   });
 
-  it('selectedMeta should receive value of meta', () => {
+  it("selectedMeta should receive value of meta", () => {
     let testMeta = (
       {
         id: 1,
@@ -54,15 +53,15 @@ describe("MetasComponent", () => {
         headshotsFront: "",
         headshotsBack: "",
         weaponry: [""],
-        level: []        
+        level: []
       }
     );
     comp.onSelect(testMeta);
     expect(comp.selectedMeta).toEqual(testMeta);
   });
 
-  it('should prepare to display the list of metas', () => {
-    let metaList = fixture.debugElement.queryAll(By.css('ul'));
+  it("should prepare to display the list of metas", () => {
+    let metaList = fixture.debugElement.queryAll(By.css("ul"));
     expect(metaList.length).toBeGreaterThan(0);
   });
